+++ plugins/wotlk/rotations/deathknight/frost/v3/01_amstaffix_deathknight_frost_loader.lua
@@ -17,23 +17,19 @@
 
 local ID = "CR>DK/F"
 local LIB_LINK = "https://raw.githubusercontent.com/AmsTaFFix/gmr-stuff/main/plugins/wotlk%2Bclassic%2Bretail/custom/00_amstlib.lua"
-local ROTATION_LINK = ""
----@type DeathKnightBloodConfig
+local ROTATION_LINK = "https://raw.githubusercontent.com/AmsTaFFix/gmr-stuff/main/plugins/wotlk/rotations/deathknight/frost/v3/02_amstaffix_deathknight_frost_rotation.lua"
+---@type DeathKnightFrostV3Config
 local Config = {
     ---Toggle debug mode. Turn on, if you encounter some issues and want to deal with it, or record a video and send
     ---to author.
-    debug = true,
+    debug = false,
     ---Use standard CombatRotation pluggable function. Change only if you know what you are doing.
     useCombatRotationLauncher = true,
     ---Use online loading feature to get last updates
-    onlineLoad = false,
-    ---Min HP to cast Rune Tap, if known
-    runeTapHpUse = 80,
-    ---Min HP to cast Vampiric Blood
-    vampiricBloodHpUse = 70,
+    onlineLoad = true,
     ---Min HP to cast Icebound Fortitude
     iceboundFortitudeHpUse = 50,
-    bloodBoilEnabled = false,
+    bloodBoilEnabled = true,
     ---Should bot use only blood runes for blood-runes offensive spells like Blood Boil, Heart/Blood Strike
     useBloodFillersWithBloodRunesOnly = true,
     ---Min enemies to start using Blood Boil instead off Blood/Hearth Strike
@@ -48,39 +44,17 @@ local Config = {
     ---Min enemies to cast raise dead spell
     minEnemiesCountToRaiseDead = 2,
 
-    ---Should use Rune Strike
-    useRuneStrike = true,
-    ---Delay between Rune Strike usage. If you encounter some issues with rune strike, tune this option.
-    useRuneStrikeDelay = 0.5,
-
-    useDeathStrike = false,
-    ---Min HP to start using Death Strike more often
-    minHPToUseDeathStrikeMoreOften = 90,
     minHPToUseDeathPact = 80,
 
-    useFrostStrike = true,
-
-    useObliterate = true,
-
-    usePlagueStrikeAsFiller = false,
-
-    useIcyTouchAsFiller = false,
+    useHowlingBlastMinEnemies = 3,
 
     useStrangulateToInterruptCasts = true,
     useDeathGripToInterruptCasts = true,
 
-    useUnholyFrenzyOnSelf = true,
-    useUnholyFrenzyMinEnemies = 3,
-
     useDeathAndDecay = false,
     useDeathAndDecayMinEnemies = 3,
 
-    useDancingRuneWeaponMinEnemies = 3,
-
-    useMarkOfBloodMinEnemies = 3,
-    useMarkOfBloodMinHP = 70,
-
-    useDeathCoilOnEnemy = false,
+    useDeathCoilOnEnemy = true,
     useDeathCoilOnEnemyMaxEnemies = 1,
 
     useCorpseExplosion = true,